+++ activesupport/lib/active_support/core_ext/enumerable.rb
@@ -16,6 +16,22 @@ module Enumerable
 
   # :startdoc:
 
+  # Calculates the minimum from the extracted elements.
+  #
+  # payments = [Payment.new(5), Payment.new(15), Payment.new(10)]
+  # payments.minimum(:price) # => 5
+  def minimum(key)
+    map(&key).min
+  end
+
+  # Calculates the maximum from the extracted elements.
+  #
+  # payments = [Payment.new(5), Payment.new(15), Payment.new(10)]
+  # payments.maximum(:price) # => 15
+  def maximum(key)
+    map(&key).max
+  end
+
   # Calculates a sum from the elements.
   #
   #  payments.sum { |p| p.price * p.tax_rate }